refactor (search): Refactor search classes.
