+++ webapi/src/main/scala/org/knora/webapi/responders/v2/search/gravsearch/types/GravsearchTypeInspectionRunner.scala
@@ -189,4 +189,5 @@ class GravsearchTypeInspectionRunner(val system: ActorSystem,
             }
         }
     }
+
 }
